+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+### Version 3.17 [23-04-2025]
+**All WoW Versions:**  
+ElvUI 13.89+ compatibility  
+Added luckyone:target Tags with and without class color suffix  
+Added luckyone:target:last Tags with and without class color suffix  
+Default font size for Pet, Player, Target, Focus is now 12 (was 14) to match other frames  
+Fixed a rare bug that could cause the BigWigs and Details installer to fail  
+Reworked Console Variables config section (Updated lists and descriptions)  
+Reworked Developer config section  
+**Retail:**  
+Updated LuckyoneUI Healer WeakAura Edits (Druid, Evoker, Monk)  
+Updated TOC for 11.1.5  
+Updated WindTools profile (Disabled keystone in tooltips)  
+**Classic:**  
+Updated TOC for 1.15.7  
+___
 ### Version 3.16 [28-03-2025]
 **All WoW Versions:**  
 ElvUI 13.87+ compatibility